requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
openai
azure-identity
pydantic
